move author and link logic to new component
